Refactor: Nextjs에서 권장하는 ESLint 규칙에 따라 코드 리팩토링 - <img> -> <Image> 태그 변경 - authOptions 분리 및 function으로 수정, 이에 따라 호출하는 곳의 import 주소 변경 - `" `-> `&quot;` 수정
