+++ src/app/api/selections/[selectionId]/reviews/[reviewId]/route.ts
@@ -1,4 +1,4 @@
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import authOptions from "@/libs/authOptions";
 import { checkIfFileExistsInS3, deleteFileFromS3, uploadFileToS3 } from "@/libs/s3";
 import { deleteSelectionReviews, extractFilePathFromUrl, getSelectionReviewImages, putSelectionReviews } from "@/services/selectionReview.services";
 import { uuidToBinary, uuidToString } from "@/utils/uuidToBinary";
@@ -13,7 +13,7 @@ export async function PUT (
     const selectionId = parseInt(params.selectionId, 10);
     const reviewId = params.reviewId;
     
-    const session = await getServerSession(authOptions);
+    const session = await getServerSession(authOptions());
     const userId = session?.user?.id;
 
     if (!userId) {
@@ -107,7 +107,7 @@ export async function DELETE (
 ) {
   try {
     const reviewId = params.reviewId;
-    const session = await getServerSession(authOptions);
+    const session = await getServerSession(authOptions());
     const userId = session?.user?.id;
 
     if (!userId) {
