+++ src/utils/authUtils.ts
@@ -3,7 +3,7 @@ import { NextRequest, NextResponse } from "next/server";
 import { UnauthorizedError } from "./errors";
 import { ErrorResponse } from "@/models/user.model";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import authOptions from "@/libs/authOptions";
 
 export const getTokenForAuthentication = async (req: NextRequest) : Promise<JWT> => {
   const token = await getToken({req});
@@ -14,7 +14,7 @@ export const getTokenForAuthentication = async (req: NextRequest) : Promise<JWT>
 };
 
 export const userIdValidator = async (userId: string): Promise<NextResponse<ErrorResponse> | null> => {
-  const session = await getServerSession(authOptions);
+  const session = await getServerSession(authOptions());
   const session_userId = session ? session.user.id : null;
   if (session_userId === null || parseInt(userId) !== session_userId) {
     return NextResponse.json({ error: "사용자 ID가 일치하지 않습니다." }, { status: 403 });
